# BrainPortal/public/swagger/cbrain-6.2.0-swagger.yaml
...
swagger: '2.0'
consumes:
  - application/json
  - application/x-www-form-urlencoded
  - multipart/form-data
info:
  title: CBRAIN API
  description: API for interacting with the CBRAIN Platform
  version: "6.2.0"
  license:
    name: GNU Public License, Version 3
    url: https://opensource.org/licenses/GPL-3.0
  contact:
    name: CBRAIN Development Team
    url: https://github.com/aces/cbrain
host: "localhost:3000"
securityDefinitions:
  BrainPortalSession:
    type: apiKey
    name: cbrain_api_token
    in: query
security:
  - BrainPortalSession: []
schemes:
  - http
basePath: /
produces:
  - application/json
  - application/xml
paths:
...
  /bourreaux:
    get:
      summary: Get a list of the Bourreaux available to be used by the current user.
      description: |
        This method returns a list of Bourreau objects.
      parameters:
        - name: page
          in: query
          description: "Page number when paginating. See also the per_page parameter"
          required: false
          type: integer
        - name: per_page
          in: query
          description: "Size of each page when paginating. See also the page parameter"
          required: false
          type: integer
      tags:
        - Bourreaux
      responses:
        '200':
          description: |
            An array of Bourreau objects describing an execution
            server available to be used by the current user.
          schema:
            type: array
            items:
              $ref: '#/definitions/Bourreau'
        '401':
          description: No session created yet.
...
  /tasks:
    get:
      summary: Get the list of Tasks.
      description: |
        This method returns the list of Tasks accessible to the current user.
      parameters:
        - name: page
          in: query
          description: "Page number when paginating. See also the per_page parameter"
          required: false
          type: integer
        - name: per_page
          in: query
          description: "Size of each page when paginating. See also the page parameter"
          required: false
          type: integer
      tags:
        - Tasks
      responses:
        '200':
          description: List of all accessible Tasks.
          schema:
            type: array
            items:
              $ref: '#/definitions/CbrainTask'
        '401':
          description: No Session created yet.
    post:
      summary: Create a new Task.
      description: |
        This method allows the creation of a new Task.
      tags:
        - Tasks
      parameters:
        - in: body
          name: cbrain_task
          description: The task to create.
          required: true
          schema:
            $ref: '#/definitions/CbrainTaskModReq'
      responses:
        '200':
          description: Task created successfully.
          schema:
            type: array
            items:
              $ref: '#/definitions/CbrainTask'
        '401':
          description: No Session created yet.
...
definitions:
...
  Bourreau:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Unique numerical ID for the bourreau.
      name:
        type: string
        description: Name given by the creator to the bourreau.
      user_id:
        type: integer
        format: int64
        description: ID of the creator of the bourreau.
      group_id:
        type: integer
        format: int64
        description: ID of the group allowed to use the bourreau.
      online:
        type: string
        format: boolean
        description: online
      read_only:
        type: string
        format: boolean
        description: Specifies whether the bourreau is read-only or can be modified.
      description:
        type: string
        description: Description of the bourreau.
...
  CbrainTask:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Unique identifier for the Task.
      type:
        type: string
        description: The task type
      user_id:
        type: integer
        format: int64
        description: ID of the User who created the Task.
      group_id:
        type: integer
        format: int64
        description: ID of the group that this task is being run in.
      bourreau_id:
        type: integer
        format: int64
        description: ID of the Bourreau the Task was launched on.
      tool_config_id:
        type: integer
        format: int64
        description: >-
          ID number that specifies which Tool Config to use. The Tool Config
          specifies environment variables and other system-specific scripts
          necessary for the Task to be run in the target environment.
      batch_id:
        type: integer
        format: int64
        description: >-
          ID of the batch this task was launched as part of. Batches of tasks
          consist of the same task, with the same parameters, being run on many
          different input files.
      params:
        type: object
        description: >-
          Parameters used as inputs to the scientific calculation associated
          with the task.
      status:
        type: string
        description: Current status of the task.
      created_at:
        type: string
        format: dateTime
        description: Date created.
      updated_at:
        type: string
        format: dateTime
        description: Last updated.
      run_number:
        type: integer
        format: int64
        description: The number of times that this task was run.
      results_data_provider_id:
        type: integer
        format: int64
        description: >-
          ID of the Data Provider that contains the Userfile that represents the
          results of the task.
      cluster_workdir_size:
        type: integer
        format: int64
        description: size of workdirectory
      workdir_archived:
        type: string
        format: boolean
        description: >-
          Boolean variable that indicates whether the working directory of the
          task is available on the processing server or has been archived and is
          no longer accessible.
      workdir_archive_userfile_id:
        type: integer
        format: int64
        description: >-
          ID of the userfile created as part of the archival process, if the
          working directory of the task has been archived.
      description:
        type: string
        description: Description of the Task.
...
  CbrainTaskModReq:
    type: object
    properties:
      cbrain_task:
        $ref: '#/definitions/CbrainTask'
...